Added shared library for test application: Svg.Tests.Common
- Refactored the shared test codes
- Removed Svg.Benchmark dependency on Svg.UnitTests
