+++ Tests/SvgW3CTestRunner/SvgW3CTestRunner.csproj
@@ -29,12 +29,11 @@
 
     <ItemGroup Condition=" '$(TargetFramework)' == 'net462'">
         <Reference Include="System.Design" />
-        <Reference Include="System.Net.Http" />
-        <Reference Include="System.IO.Compression" />
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\..\Source\Svg.csproj" />
+        <ProjectReference Include="..\Svg.Tests.Common\Svg.Tests.Common.csproj" />
     </ItemGroup>
 
     <ItemGroup>
